feat(web): bind Linq group owner from add actor
